Working to replace connection string
There is a lot of garbled code that can be removed but I finally got the
script searching the entire file storing the connection string as a
variable and replacing it while leaving the backup copy untouched.

I use the backup copy to read from and the original to write to.